Update ContextMock in tests
